+++ netatmo-get-measurements.html
@@ -33,7 +33,7 @@
         outputs:1,
         icon: "font-awesome/fa-cloud",
         label: function() {
-            return "get "+this.types+' measurements \n over '+this.scale||"get measurements";
+            return this.name || "get "+this.types+' measurements over '+this.scale||"get measurements";
         },
         oneditprepare: function() {
             oneditprepareDeviceId(true); // TODO: show only allowed devices
@@ -58,12 +58,12 @@
             </a>
         </div>
 	</div>
-    <div class="form-tips"><b>Tip:</b> Can be left blank to get all your devices.</div>
-    <br/>
     <div class="form-row">
         <label for="node-input-moduleId"><i class="fa fa-home"></i> Module ID</label>
         <input type="text" id="node-input-moduleId" placeholder="70:ee:.....">
     </div>
+    <div class="form-tips"><b>Tip:</b> Can be left blank to get all your devices.</div>
+    <br/>
     <div class="form-row">
         <label for="node-input-scale"><i class="fa fa-clock"></i> Scale</label>
         <select id="node-input-scale" value="max">
@@ -98,20 +98,35 @@
 </script>
 
 <script type="text/html" data-help-name="get measurements">
-   <p>This node returns measurements from a device or a module</p>
+    <p>This node returns (historical) measurements from a device or a module</p>
+    <p>Devices:
 
-   <h3>Inputs</h3>
+https://dev.netatmo.com/apidocumentation/control#getmeasure
+read_magellan for Legrand Wiring device or Electrical panel products (consumption data)
+read_smarther for Smarther with Netatmo thermostat (boiler/cooler status)
+read_thermostat for Smart Modulating thermostat (boiler status)
 
-   <dl class="message-properties">
-       <dt class="optional">msg.deviceId<span class="property-type">string</span>
-       </dt>
-       <dd>If specified the function will overwrite the node's Device ID.</dd>
-       <dt class="optional">msg.getFavorites<span class="property-type">bool</span>
-       </dt>
-       <dd>If specified the function will overwrite the node's Get Favorites property.</dd>
-   </dl>
+https://dev.netatmo.com/apidocumentation/energy#getmeasure
 
-   <h3>Outputs</h3>
+NATherm1 for read_thermostat Smart Thermostat
+
+https://dev.netatmo.com/apidocumentation/weather#getmeasure
+
+read_station for Smart Home Weather Station
+
+    </p>
+    <h3>Inputs</h3>
+
+    <dl class="message-properties">
+        <dt class="optional">msg.deviceId<span class="property-type">string</span>
+        </dt>
+        <dd>If specified the function will overwrite the node's Device ID.</dd>
+        <dt class="optional">msg.getFavorites<span class="property-type">bool</span>
+        </dt>
+        <dd>If specified the function will overwrite the node's Get Favorites property.</dd>
+    </dl>
+
+    <h3>Outputs</h3>
     <dl class="message-properties">
         <dt>payload<span class="property-type">(object)</span>
         </dt>
