+++ netatmoNodes.js
@@ -131,203 +131,10 @@ module.exports = function(RED) {
     RED.nodes.registerType("get home data",NetatmoGetHomeData);
    /***************************************************************/
 
-    function NetatmoHomesData(config) {
-
-        RED.nodes.createNode(this,config);
-        this.creds = RED.nodes.getNode(config.creds);
-        var node = this;
-        this.on('input', function(msg) {
-            this.homeId = msg.homeId || config.homeId || '';
-            this.gatewayTypes = msg.gatewayTypes || config.gatewayTypes || '';
-			
-            var netatmo = require('netatmo');
-
-            var auth = {
-                "client_id": this.creds.client_id,
-                "client_secret": this.creds.client_secret,
-                "username": this.creds.username, 
-                "password": this.creds.password
-            };
-            var api = new netatmo(auth);
-            
-            api.on("error", function(error) {
-                node.error(error);
-            });
-
-            api.on("warning", function(error) {
-                node.warn(error);
-            });                 
-            
-            var options = {
-            };
-
-            // Optional
-            if (this.homeId !== '') {
-                options.home_id = this.homeId;
-            }
-
-            if (this.gatewayTypes !== '') {
-                options.gateway_types = this.gatewayTypes;
-            }
-			
-            api.homesData(options,function(err, body) {
-                msg.payload = body;
-                node.send(msg);
-            });
-        });
-
-    }
-    RED.nodes.registerType("homes data",NetatmoHomesData);
 
     /***************************************************************/
-	function NetatmoHomeStatus(config) {
-
-        RED.nodes.createNode(this,config);
-        RED.nodes.getNode(config.creds);
-
-        var node = this;
-        this.on('input', function(msg) {
-            this.homeId = msg.homeId || config.homeId || '';
-            this.deviceTypes = msg.deviceTypes || config.deviceTypes || '';
-			
-            var netatmo = require('netatmo');
-
-            const auth = {
-                "client_id": this.credentials.client_id,
-                "client_secret": this.credentials.client_secret,
-                "username": this.credentials.username, 
-                "password": this.credentials.password
-            };
-            var api = new netatmo(auth);
-            
-            api.on("error", function(error) {
-                node.error(error);
-            });
-
-            api.on("warning", function(error) {
-                node.warn(error);
-            });                 
-            
-            var options = {
-				'home_id' : this.homeId,
-            };
-
-            // Optional
-            if (this.deviceTypes !== ''){
-                options.deviceTypes = this.deviceTypes;
-            }
-			
-            api.homeStatus(options,function(err, body) {
-                msg.payload = body;
-                node.send(msg);
-            });
-        });
-
-    }
-    RED.nodes.registerType("home status",NetatmoHomeStatus);
     /***************************************************************/
-    function NetatmoSetThermMode(config) {
-
-        RED.nodes.createNode(this,config);
-        this.creds = RED.nodes.getNode(config.creds);
-        var node = this;
-        this.on('input', function(msg) {
-            this.homeId = msg.homeId || config.homeId || '';
-            this.mode = msg.mode || config.mode || '';
-            this.endtime = msg.endtime || config.endtime || '';
-			
-            var netatmo = require('netatmo');
-
-            var auth = {
-                "client_id": this.creds.client_id,
-                "client_secret": this.creds.client_secret,
-                "username": this.creds.username, 
-                "password": this.creds.password
-            };
-            var api = new netatmo(auth);
-            
-            api.on("error", function(error) {
-                node.error(error);
-            });
-
-            api.on("warning", function(error) {
-                node.warn(error);
-            });                 
-            
-            var options = {
-                home_id : this.homeId,
-                mode : this.mode,
-            };
-
-            if ((this.endtime !== '')&&(this.endtime !== null)){
-                if(this.endtime > 1552236804) { // Correction if we got a UNIX Tmestamp
-                  options.endtime = Math.floor(this.endtime / 1000);
-                }
-                else {
-                  options.endtime = this.endtime;
-                }
-            }
-			
-            api.setThermMode(options);
-        });
-
-    }
-    RED.nodes.registerType("set therm mode",NetatmoSetThermMode);
     /***************************************************************/
-    function NetatmoSetRoomThermPoint(config) {
-
-        RED.nodes.createNode(this,config);
-        this.creds = RED.nodes.getNode(config.creds);
-        var node = this;
-        this.on('input', function(msg) {
-            this.homeId = msg.homeId || config.homeId || '';
-            this.roomId = msg.roomId || config.roomId || '';
-            this.mode = msg.mode || config.mode || '';
-            this.temp = msg.temp || config.temp || '';
-            this.endtime = msg.endtime || config.endtime || '';
-			
-            var netatmo = require('netatmo');
-
-            var auth = {
-                "client_id": this.creds.client_id,
-                "client_secret": this.creds.client_secret,
-                "username": this.creds.username, 
-                "password": this.creds.password
-            };
-            var api = new netatmo(auth);
-            
-            api.on("error", function(error) {
-                node.error(error);
-            });
-
-            api.on("warning", function(error) {
-                node.warn(error);
-            });                 
-            
-            var options = {
-                home_id : this.homeId,
-                room_id : this.roomId,
-                mode : this.mode,
-            };
-
-            if ((this.temp !== '')&&(this.temp !== null)){
-                options.temp = this.temp;
-            }
-
-            if ((this.endtime !== '')&&(this.endtime !== null)){
-                if(this.endtime > 1552236804) { // Correction if we got a UNIX Tmestamp
-                    options.endtime = Math.floor(this.endtime / 1000);
-                  }
-                  else {
-                    options.endtime = this.endtime;
-                  }
-            }
-			
-            api.setRoomThermPoint(options);
-        });
-
-    }
-    RED.nodes.registerType("set room therm point",NetatmoSetRoomThermPoint);
     /***************************************************************/
     function NetatmoGetPublicData(config) {
         RED.nodes.createNode(this,config);
